Extract control fitting into its own module
z_score.py held the moving-average filter, three overlapping control-fitting
paths, the dF/F normalization and the z-scoring. Control fitting moves to
control_fit.py, taking filterSignal, deltaFF and the chunk-length validation
with it (all three exist only to serve the fit), and leaving z_score.py with
the chunk orchestration and the z-score itself.

The three estimation paths collapse into one. A control fit has two
independent choices -- which regressors (control alone, or control plus the
photobleaching term) and which samples estimate the coefficients (the segment
being corrected, or a baseline epoch) -- but they were a bool flag and a
pre-computed tuple that bypassed the fitting branch, so they did not compose.
Both are now arguments to estimate_control_fit, which returns a
ControlFitModel that apply_control_fit evaluates over any segment.

That makes baseline-epoch fitting work with the photobleaching term instead of
being rejected by it, so the guard forbidding the combination is gone.
Baseline-epoch fitting is what produces a slow drift in the first place -- the
frozen coefficients diverge from the signal as the two channels bleach at
different rates -- so it is the mode that most needs the term.

The decay constant is now found by a bounded search over its logarithm rather
than a grid, and its bounds are loose enough not to constrain the result.
